% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{A}
\alias{A}
\title{Compute Availability}
\usage{
A(x, k1, k2, s1, output_file = NULL, plot = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{k1}{A numeric, start position}

\item{k2}{A numeric, end position}

\item{s1}{Character vector of the subspace working states among the state space vector s.t. s1<|s|}

\item{output_file}{A file containing matrix of availability probabilities}
}
\value{
A matrix with Availability score at each position
}
\description{
Compute Availability
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda,1,1,c("a","c","g","t"))
k1 <- 1
k2 <- 200
s1 <- c("c","t")  # vector of working states
AVA.out <- "C:\\\\...\\\\file.txt"
A(dmm,k1,k2,s1, output_file=AVA.out,plot=FALSE)
}
\author{
Alexandre Seiller
}
