% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{bic.dmmsum}
\alias{bic.dmmsum}
\title{Evaluate BIC}
\usage{
\method{bic}{dmmsum}(x, sequences)
}
\arguments{
\item{x}{An object of class "dmm", \link[drimmR]{dmmsum}}

\item{sequences}{A character vector or a list of character vector representing the sequence}
}
\value{
A numeric, BIC
}
\description{
Evaluate BIC
}
\examples{
data(lambda, package = "drimmR")
sequence <- c("a","g","g","t","c","g","a","t","a","a","a")
dmm<-dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
bic(dmm,sequence)
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{loglik}, \link[drimmR]{bic}
}
\author{
Victor Mataigne, Alexandre Seiller
}
