% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{Distribution_evol}
\alias{Distribution_evol}
\title{Distributions for a range of positions}
\usage{
Distribution_evol(
  x,
  start = 1,
  end = NULL,
  step = NULL,
  output_file = NULL,
  plot = FALSE
)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{start}{Start position}

\item{end}{End position}

\item{output_file}{A file containing matrix of distributions}

\item{plot}{FALSE (no figure plot of dist evolution); TRUE (figure plot)}
}
\value{
A matrix of distributions with position and probability of states
}
\description{
Distributions for a range of positions
}
\examples{
#' data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
DIST.out <- "C:\\\\...\\\\file.txt"
Distribution_evol(dmm,start=1,end=length(lambda)-1,step=10000, output_file=DIST.out, plot=FALSE)
}
\author{
Alexandre Seiller
}
