% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drimmR-package.R
\docType{package}
\name{drimmR-package}
\alias{drimmR}
\alias{drimmR-package}
\title{drimmR-package}
\description{
This package performs the estimation, simulation and the exact computation of associated
reliability measures of drifting Markov models (DMMs). These are particular non-homogeneous Markov chains
for which the Markov transition matrix is a linear (polynomial) function of two (several) Markov transition matrices.
Several statistical frameworks are taken into account (one or several samples, complete or incomplete samples, models
of the same length). Computation of probabilities of appearance of a word along a sequence
under a given model are also considered.
}
