% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{getDistribution}
\alias{getDistribution}
\title{Get distribution at a given position or at every positions}
\usage{
getDistribution(x, pos, all.pos = FALSE, internal = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{pos}{An integer, a position}

\item{all.pos}{FALSE (evaluation at pos index) ; TRUE (evaluation for all pos index)}

\item{internal}{FALSE (default) ; TRUE (for internal use of distrib_evol function)}
}
\value{
distribution at position x or at every positions
}
\description{
Get distribution at a given position or at every positions
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
t <- 10
getDistribution(dmm,t)
}
\author{
Alexandre Seiller
}
