% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{getStationaryLaw.dmmsum}
\alias{getStationaryLaw.dmmsum}
\title{Get stationary law}
\usage{
\method{getStationaryLaw}{dmmsum}(x, pos, all.pos = FALSE, internal = FALSE)
}
\arguments{
\item{x}{An object of class "dmm", \link[drimmR]{dmmsum}}

\item{pos}{An integer for position}

\item{all.pos}{FALSE (default, evaluation at position index) ; TRUE (evaluation for all position indices)}

\item{internal}{FALSE (default) ; TRUE (for internal use of dmmsum initial law and word applications)}
}
\value{
A vector or matrix of stationary law probabilities
}
\description{
Evaluate stationary law at a given position or at every position
}
\details{
Stationary law at position t is evaluated by solving \eqn{\mu_t \ \pi_{\frac{t}{n}} = \mu}
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
t <- 10
getStationaryLaw(dmm,pos=t)

}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{stationaryLaw_evol}, \link[drimmR]{getDistribution}
}
\author{
Alexandre Seiller
}
