% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{getTransitionMatrix}
\alias{getTransitionMatrix}
\title{Get transition S3 generic function}
\usage{
getTransitionMatrix(x, pos)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{pos}{An integer, a position}
}
\value{
A transition matrix at a given position
}
\description{
Get transition S3 generic function
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\author{
Victor Mataigne, Alexandre Seiller
}
