% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lambda}
\alias{lambda}
\title{lambda genome}
\format{
A vector object of class \code{"Rdata"}.
}
\usage{
data(lambda)
}
\description{
Complete data from phage genome [WT71] of length 48502
}
\examples{
data(lambda)
}
\keyword{datasets}
