% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{length_probas}
\alias{length_probas}
\title{Probability of appearance of the Observed word of size n in a sequence at several positions}
\usage{
length_probas(n, sequence, pos, x, output_file = NULL, plot = FALSE)
}
\arguments{
\item{n}{An integer, the length word}

\item{sequence}{A vector of characters}

\item{pos}{A vector of positions}

\item{x}{An object of class "dmm"}
}
\value{
a dataframe
}
\description{
Probability of appearance of the Observed word of size n in a sequence at several positions
}
\examples{
length_probas(2, sequence, c(100,200), x)
}
\author{
Victor Mataigne, Alexandre Seiller
}
