% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{loglik.dmmsum}
\alias{loglik.dmmsum}
\title{Evaluate Log-likelihood}
\usage{
\method{loglik}{dmmsum}(x, sequences)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{sequences}{A character vector or a list of character vectors representing the sequence}
}
\value{
A list of log-likelihood (numeric)
}
\description{
Evaluate Log-likelihood
}
\examples{
data(lambda, package = "drimmR")
sequence <- c("a","g","g","t","c","g","a","t","a","a","a")
dmm <-dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
loglik(dmm,sequence)

}
\author{
Annthomy Gilles, Alexandre Seiller
}
