% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood}
\usage{
loglik(x, sequences)
}
\arguments{
\item{x}{An object for which the log-likelihood of the DMM can be computed.}

\item{sequences}{A vector of character or list of vectors representing the sequences for which the
log-likelihood of the model must be computed.}
}
\value{
A list of loglikelihood (numeric)
}
\description{
Generic function computing the log-likelihood of the model \code{x},
  with the list of sequences \code{sequences}.
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\author{
Annthomy Gilles, Alexandre Seiller
}
