% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{M}
\alias{M}
\title{Evaluate Maintainability}
\usage{
M(x, k1, k2, s1, output_file = NULL, plot = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{k1}{start position (numeric)}

\item{k2}{end position (numeric)}

\item{s1}{Character vector of the subspace working states among the state space vector s.t. s1 < s}

\item{output_file}{A file containing matrix of maintainability probabilities}

\item{plot}{FALSE (no figure plot of maintainability by position); TRUE (figure plot)}
}
\value{
A matrix with Maintainability score at each position
}
\description{
Estimation of maintainability for (ergodic) repairable system at time \eqn{k \in N}.
}
\details{
The maintainability at time \eqn{k \in N} of a repairable system is the probability that the system is repaired up to time \eqn{l},
given that is has failed at time \eqn{l=0}.
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda,1,1,c("a","c","g","t"))
k1 <- 1
k2 <- 200
s1 <- c("c","t")  # vector of working states
MAIN.out <- "C:\\\\...\\\\file.txt"
M(dmm,k1,k2,s1, output_file=MAIN.out,plot=FALSE)

}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}
}
\author{
Alexandre Seiller
}
