% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{maintainability}
\alias{maintainability}
\title{Maintainability function}
\usage{
maintainability(x, k1, k2, upstates, output_file = NULL, plot = FALSE)
}
\arguments{
\item{x}{An object of class \code{dmm}}

\item{k1}{Start position :  a positive integer giving the start position along the sequence from which the maintainabilities of the DMM should be computed}

\item{k2}{End position :  a positive integer giving the end position along the sequence until which the maintainabilities of the DMM should be computed}

\item{upstates}{Character vector of the subspace working states among the state space vector such that upstates < s}

\item{output_file}{(Optional) A file containing matrix of maintainability probabilities (e.g, "C:/.../MAIN.txt")}

\item{plot}{\code{FALSE} (no figure plot of maintainability by position); \code{TRUE} (figure plot)}
}
\value{
A matrix with positions and maintainability probabilities of states (and figure plot)
}
\description{
Maintainability of a system at time \eqn{k \in N}.
}
\details{
The maintainability at time \eqn{k \in N} of a system is the probability that the system is repaired up to time \eqn{l},
given that is has failed at time \eqn{l=0}.
}
\examples{

data(lambda, package = "drimmR")
dmm <- fitdmm(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq", fit.method="sum")
k1 <- 1
k2 <- 200
upstates <- c("c","t")  # vector of working states
maintainability(dmm,k1,k2,upstates,plot=TRUE)

}
\references{
\insertRef{BaVe2018}{drimmR}
}
\seealso{
\link[drimmR]{fitdmm}, \link[drimmR]{getTransitionMatrix}
}
\author{
Alexandre Seiller
}
