% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{R}
\alias{R}
\title{Evaluate Reliability}
\usage{
R(x, k1, k2, s1, output_file = NULL, plot = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{k1}{A numeric, start position}

\item{k2}{A numeric, end position}

\item{s1}{Character vector of the subspace working states among the state space vector s.t. s1 < s}

\item{output_file}{A file containing matrix of reliability probabilities}

\item{plot}{FALSE (no figure plot of reliability by position); TRUE (figure plot)}
}
\value{
A matrix with Reliability score at each position
}
\description{
Estimation of reliability for (ergodic) repairable or (non-ergodic) non-repairable systems at time \eqn{l \in N}
}
\details{
The reliability at time \eqn{l \in N} is the probability thaht the system has functioned without failure in the period \eqn{[0, l]}
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda,1,1,c("a","c","g","t"))
k1 <- 1
k2 <- 200
s1 <- c("c","t")  # vector of working states
REL.out <- "C:\\\\...\\\\file.txt"
R(dmm,k1,k2,s1, output_file=REL.out,plot=FALSE)

}
\seealso{
\code{\link[drimmR]{dmmsum}}
}
\author{
Alexandre Seiller
}
