% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{reliability}
\alias{reliability}
\title{Reliability function}
\usage{
reliability(x, k1, k2, upstates, output_file = NULL, plot = FALSE)
}
\arguments{
\item{x}{An object of class \code{dmm}}

\item{k1}{Start position :  a positive integer giving the start position along the sequence from which the reliabilities of the DMM should be computed}

\item{k2}{End position :  a positive integer giving the end position along the sequence until which the reliabilities of the DMM should be computed}

\item{upstates}{Character vector of the subspace working states among the state space vector such that upstates < s}

\item{output_file}{(Optional) A file containing matrix of reliability probabilities (e.g, "C:/.../REL.txt")}

\item{plot}{\code{FALSE} (no figure plot of reliability by position); \code{TRUE} (figure plot)}
}
\value{
A matrix with positions and reliability probabilities of states (and figure plot)
}
\description{
Reliability or the survival function of a system at time \eqn{l \in N}
}
\details{
The reliability at time \eqn{l \in N} is the probability that the system has functioned without failure in the period \eqn{[0, l]}
}
\examples{

data(lambda, package = "drimmR")
dmm <- fitdmm(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq", fit.method="sum")
k1 <- 1
k2 <- 200
upstates <- c("c","t")  # vector of working states
reliability(dmm,k1,k2,upstates,plot=TRUE)


}
\references{
\insertRef{BaVe2018}{drimmR}
}
\seealso{
\link[drimmR]{fitdmm}, \link[drimmR]{getTransitionMatrix}
}
\author{
Alexandre Seiller
}
