% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{simulate.dmmsum}
\alias{simulate.dmmsum}
\title{Simulate a sequence with the Drifting Markov Model}
\usage{
\method{simulate}{dmmsum}(x, output_file, model_size = NULL)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{output_file}{directory path of the output file.fa}

\item{model_size}{Size of the model}
}
\description{
Simulate a sequence with the Drifting Markov Model
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
simulate(dmm,"simulated_data.out",20000)

}
\author{
Annthomy Gilles, Alexandre Seiller
}
