% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate a sequence with the Drifting Markov Model}
\usage{
simulate(x, output_file, model_size = 1e+05)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{output_file}{A file containing matrix of probabilities}

\item{model_size}{Size of the model}
}
\description{
Simulate a sequence with the Drifting Markov Model
}
\examples{
data("sequence_example")
dmm <- point_estimate(sequence_example, 2, 1, c('a','c','g','t'), 1000000)
simulation(dmm,"simulated_data.out",20000)
}
\author{
GILLES Anthomy
}
