% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word_expect}
\alias{word_expect}
\title{Expectation of a word in a DMM}
\usage{
word_expect(word, x)
}
\arguments{
\item{word}{A subsequence}

\item{DMM}{An object of class "dmm"}
}
\value{
a float
}
\description{
Expectation of a word in a DMM
}
\examples{
word_expect("atcggatc", DMM)
}
\author{
Victor Mataigne
}
