% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word_proba}
\alias{word_proba}
\title{Probability of a word at a position t of a DMM}
\usage{
word_proba(word, t, x, output_file = NULL)
}
\arguments{
\item{word}{A subsequence}

\item{t}{A position}

\item{x}{An object of class "dmm"}
}
\value{
A numeric, probability of \code{word}
}
\description{
Probability of a word at a position t of a DMM
}
\examples{
word_proba("aggctga",4,x)
}
\author{
Victor Mataigne
}
