% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word_probas}
\alias{word_probas}
\title{Probabilities of a word at several positions of a DMM}
\usage{
word_probas(word, t, x, output_file = NULL, plot = FALSE)
}
\arguments{
\item{word}{A subsequence}

\item{t}{A vector of positions}

\item{x}{An object of class "dmm"}

\item{output_file}{A file containing the vector of probabilities}

\item{plot}{FALSE (no figure plot of word probabilities); TRUE (figure plot)}
}
\value{
A numeric vector, probabilities of \code{word}
}
\description{
Probabilities of a word at several positions of a DMM
}
\examples{
#' data(lambda, package = "drimmR")
mod <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
PROB.out <- "C:\\\\...\\\\file.txt"
word_probas("aggctga",c(100,300),x, output_file=PROB.out, plot=FALSE)
}
\author{
Victor Mataigne, Alexandre Seiller
}
