% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{words_of_size_n_probas}
\alias{words_of_size_n_probas}
\title{Probability of words of a given size at several positions of a DMM}
\usage{
words_of_size_n_probas(n, pos, x, output_file = NULL)
}
\arguments{
\item{n}{An integer, the length word}

\item{pos}{A vector of positions}

\item{x}{An object of class "dmm"}
}
\value{
a dataframe
}
\description{
Probability of words of a given size at several positions of a DMM
}
\examples{
words_of_size_n_probas(2, c(100,200), DMM)
}
\author{
Victor Mataigne
}
