% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{words_probabilities}
\alias{words_probabilities}
\title{Probability of appearance of several words at several positions of a DMM}
\usage{
words_probabilities(words, pos, x, output_file = NULL)
}
\arguments{
\item{words}{A vector of characters containing words}

\item{pos}{A vector of integer positions}

\item{x}{An object of class \code{dmm}}

\item{output_file}{(Optional) A file containing the matrix of probabilities (e.g,"C:/.../PROB.txt")}
}
\value{
A dataframe of word probabilities along the positions of the sequence
}
\description{
Probability of appearance of several words at several positions of a DMM
}
\examples{

data(lambda, package = "drimmR")
dmm <- fitdmm(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq", fit.method="sum")
words <- c("atcgattc", "taggct", "ggatcgg")
pos <- c(100,300)
words_probas <- words_probabilities(words=words,pos=pos,dmm)


# Figure plot example :
library(ggplot2)
library(reshape2)

colnames(words_probas) <- c("position",words[1:length(words)])
words_probas<- subset(words_probas, position \%in\% c(pos[1]:pos[2]))
# warning is ok :
frame <- data.frame(reshape2::melt(words_probas[,-1]))
frame <- cbind(rep(c(pos[1]:pos[2]),length(words)),frame)
colnames(frame) <-c("position","word","probability")
fig <- list()
for (o in c(1:length(words))){

  fig[[o]] <- frame \%>\% dplyr::filter(word==words[o]) \%>\%  ggplot2::ggplot(aes(x=position, y=probability)) +
  ggplot2::geom_line() + geom_point() + scale_x_continuous(name= "Position",breaks = c(pos[1],pos[2])) +
    ggplot2::ggtitle(paste0("Overall frame : \n Probability of the word '", words[o],"' at each position of the frame")) +
    ggplot2::theme_bw()
}
fig

}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\seealso{
\link[drimmR]{fitdmm}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{word_probability}, \link[drimmR]{word_probabilities}
}
\author{
Victor Mataigne, Alexandre Seiller
}
