% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{words_probas}
\alias{words_probas}
\title{Probability of several words at several positions of a DMM}
\usage{
words_probas(words, pos, x, output_file = NULL, plot = FALSE)
}
\arguments{
\item{words}{A vector of characters containing words}

\item{pos}{A vector of integer positions}

\item{x}{An object of class `dmm`}

\item{output_file}{(Optional) A file containing the matrix of probabilities (e.g,"C:/.../PROB.txt")}

\item{plot}{`FALSE` (no figure plot of words probabilities); `TRUE` (figure plot)}
}
\value{
A dataframe of word probabilities along the positions of the sequence
}
\description{
Probability of several words at several positions of a DMM
}
\examples{
\dontrun{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
words_probas(c("atcgattc", "taggct", "ggatcgg"),c(100,300),dmm, plot=TRUE)
}
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{word_proba}, \link[drimmR]{word_probas}
}
\author{
Victor Mataigne, Alexandre Seiller
}
