% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{aic.dmmsum}
\alias{aic.dmmsum}
\title{Evaluate AIC}
\usage{
\method{aic}{dmmsum}(x, sequences)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{sequence}{A character vector or a list of character vector representing the sequence}
}
\value{
A list of numeric, AIC
}
\description{
Evaluate AIC
}
\examples{
data(lambda, package = "drimmR")
sequence <- c("a","g","g","t","c","g","a","t","a","a","a")
dmm <-point_estimate(lambda, 1, 1, c('a','c','g','t'), 1000000)
aic(dmm,sequence)
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{loglik}, \link[drimmR]{aic}
}
\author{
Victor Mataigne, Alexandre Seiller
}
