% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion (BIC)}
\usage{
bic(x, sequences)
}
\arguments{
\item{x}{An object for which BIC of the DMM can be computed.}

\item{sequences}{A vector or a list of vector of character representing the sequences for which the
BIC criterion must be computed.}
}
\value{
A list of BIC (numeric)
}
\description{
Generic function computing the Bayesian Information Criterion
  of the model \code{x}, with the list of sequences \code{sequences}.
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\author{
Victor Mataigne, Alexandre Seiller
}
