% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion (BIC)}
\usage{
bic(x, sequence)
}
\arguments{
\item{x}{An object of class "dmm" for which the logikelihood can be computed.}

\item{sequence}{A vector or a list of vector of character representing the sequences for which the
BIC criterion must be computed.}
}
\value{
A numeric value giving the value of the BIC.
}
\description{
Generic function computing the Bayesian Information Criterion
  of the model `x`, with the list of sequences `sequences`.
}
\author{
Victor Mataigne, Alexandre Seiller
}
