% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{failureRate}
\alias{failureRate}
\title{Evaluate failure rates}
\usage{
failureRate(
  x,
  k1,
  k2,
  s1,
  error.rate = c("BMP", "RG"),
  output_file = NULL,
  plot = FALSE
)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{k1}{start position (numeric)}

\item{k2}{end position (numeric)}

\item{s1}{Character vector of the subspace working states among the state space vector s.t. s1 < s}

\item{error.rate}{Default="BMP", then BMP-failure-rate is the method used to estimate the error rate. If error.rate= "RG",
then RG-failure rate is the method used to estimate the error rate.}

\item{output_file}{(Optional) A file containing matrix of error rates at each position (e.g, "C:/.../ER.txt")}

\item{plot}{FALSE (no figure plot of error rates by position); TRUE (figure plot)}
}
\value{
A matrix with error rate score at each position
}
\description{
Estimation of two different definition of the failure rate : the BMP-failure rate and RG-failure rate
}
\details{
The BMP-failure rate denoted by \eqn{\lambda(l), l \in N} is usually considered for
continuous time systems.

The RG-failure rate denoted by \eqn{r(l), l \in N} is adapted to work in discrete time systems.
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
k1 <- 1
k2 <- 200
s1 <- c("c","t")  # vector of working states
failureRate(dmm,k1,k2,s1,error.rate="BMP",plot=TRUE)

}
\references{
\insertRef{BaVe2018}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{reliability}
}
\author{
Alexandre Seiller
}
