% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{getDistribution.dmmsum}
\alias{getDistribution.dmmsum}
\title{Get distribution}
\usage{
\method{getDistribution}{dmmsum}(x, pos, all.pos = FALSE, internal = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{pos}{An integer for position}

\item{all.pos}{FALSE (evaluation at pos index) ; TRUE (evaluation for all pos index)}

\item{internal}{FALSE (default) ; TRUE (for internal use of distrib_evol function)}
}
\value{
A matrix of probabilities
}
\description{
Get distribution
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
t <- 10
getDistribution(dmm,t)

}
\author{
Alexandre Seiller
}
