% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{getDistribution}
\alias{getDistribution}
\title{Get distribution S3 generic function}
\usage{
getDistribution(x, pos, all.pos = FALSE, internal = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{pos}{An integer, a position}

\item{all.pos}{FALSE (evaluation at pos index) ; TRUE (evaluation for all pos indices)}

\item{internal}{FALSE (default) ; TRUE (for internal use of distrib_evol function)}
}
\value{
A vector or matrix of distribution probabilities
}
\description{
Evaluate distribution at a given position or at every position
}
\details{
Distribution at position l is evaluated by \eqn{\mu_{l} =\mu_0 \prod_{t=k}^{l} \ \pi_{\frac{t}{n}}}, \eqn{\forall l \ge k, k \in N^*} order of the DMM
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
t <- 10
getDistribution(dmm,pos=t)
}
\author{
Alexandre Seiller
}
