% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{getStationaryLaw.dmmsum}
\alias{getStationaryLaw.dmmsum}
\title{Get stationary law}
\usage{
\method{getStationaryLaw}{dmmsum}(x, pos, all.pos = FALSE, internal = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{pos}{An integer for position}

\item{all.pos}{FALSE (evaluation at pos index) ; TRUE (evaluation for all pos index)}

\item{internal}{FALSE (default) ; TRUE (for internal use of dmmsum initial law and word applications)}
}
\value{
A vector or matrix of stationary laws
}
\description{
Get stationary law
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
t <- 10
getStationaryLaw(dmm,pos=t)

}
\author{
Alexandre Seiller
}
