% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{getStationaryLaw}
\alias{getStationaryLaw}
\title{Get stationary law S3 generic function}
\usage{
getStationaryLaw(x, pos, all.pos = FALSE, internal = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{pos}{An integer, a position}

\item{all.pos}{FALSE (default, evaluation at position index) ; TRUE (evaluation for all position indices)}

\item{internal}{FALSE (default) ; TRUE (for internal use of dmmsum initial law)}
}
\value{
A vector or matrix of stationary law probabilities
}
\description{
Evaluate stationary law at a given position or at every position
}
\details{
Stationary law at position t is evaluated by solving \eqn{\mu_t \ \pi_{\frac{t}{n}} = \mu}
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\author{
Alexandre Seiller
}
