% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{getStationaryLaw}
\alias{getStationaryLaw}
\title{Get the stationary laws of the DMM}
\usage{
getStationaryLaw(x, pos, all.pos = FALSE, internal = FALSE)
}
\arguments{
\item{x}{An object of class `dmm`}

\item{pos}{A positive integer giving the position along the sequence on which the stationary law of the DMM should be computed}

\item{all.pos}{`FALSE` (default, evaluation at position index) ; `TRUE` (evaluation for all position indices)}

\item{internal}{`FALSE` (default) ; `TRUE` (for internal use of the initial law computation)}
}
\value{
A vector or matrix of stationary law probabilities
}
\description{
Generic function evaluating the stationary law of the DMM at a given position or at every position
}
\details{
\code{Stationary law} at position t is evaluated by solving \eqn{\mu_t \ \pi_{\frac{t}{n}} = \mu}
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\author{
Alexandre Seiller
}
