% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{getStationaryLaw}
\alias{getStationaryLaw}
\title{Get stationary law at a given position or at every positions}
\usage{
getStationaryLaw(x, pos, all.pos = FALSE, internal = FALSE)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{pos}{An integer, a position}

\item{all.pos}{FALSE (default, evaluation at position index) ; TRUE (evaluation for all position indices)}

\item{internal}{FALSE (default) ; TRUE (for internal use of dmmsum initial law)}
}
\value{
A vector or matrix of stationary law probabilities
}
\description{
Evaluate stationary law at a given position or at every position
}
\details{
Stationary law at position l is evaluated by solving \eqn{\mu_0 \ \pi_l = \mu}
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
t <- 10
getStationaryLaw(dmm,pos=t)
}
\author{
Alexandre Seiller
}
