% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{getTransitionMatrix}
\alias{getTransitionMatrix}
\title{Get transition S3 generic function}
\usage{
getTransitionMatrix(x, pos)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{pos}{An integer, a position}
}
\value{
A transition matrix at a given position
}
\description{
Get transition S3 generic function
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim="freq")
t <- 10
getTransitionMatrix(dmm,pos=t)
}
\author{
Victor Mataigne, Alexandre Seiller
}
