% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{getTransitionMatrix}
\alias{getTransitionMatrix}
\title{Transition matrix of the drifting Markov Model}
\usage{
getTransitionMatrix(x, pos)
}
\arguments{
\item{x}{An object for which the transition matrices of the DMM can be computed.}

\item{pos}{A positive integer giving the position along the sequence on which the transition matrix of the DMM should be computed}
}
\value{
A transition matrix at a given position
}
\description{
Generic function evaluating the transition matrix of a model \code{x} at a given position \code{pos}
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\author{
Victor Mataigne, Alexandre Seiller
}
