% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{length_probas}
\alias{length_probas}
\title{Probability of appearance of the Observed word of size n in a sequence at several positions}
\usage{
length_probas(n, sequence, pos, x, output_file = NULL, plot = FALSE)
}
\arguments{
\item{n}{An integer, the length word}

\item{sequence}{A vector of characters}

\item{pos}{A vector of integer positions}

\item{x}{An object of class "dmm"}

\item{output_file}{A file containing the vector of probabilities}

\item{plot}{FALSE (no figure plot of words probabilities); TRUE (figure plot)}
}
\value{
a dataframe of probability by position (and probability plots)
}
\description{
Probability of appearance of the Observed word of size n in a sequence at several positions
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
PROB.out <- "C:\\\\...\\\\file.txt"
length_probas(2, lambda, c(100,200), dmm, output_file=PROB.out)
}
\author{
Victor Mataigne, Alexandre Seiller
}
