% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{length_probas}
\alias{length_probas}
\title{Probability of occurrence of the observed word of size n in a sequence at several positions}
\usage{
length_probas(n, sequence, pos, x, output_file = NULL, plot = FALSE)
}
\arguments{
\item{n}{An integer, the length word}

\item{sequence}{A vector of characters}

\item{pos}{A vector of integer positions}

\item{x}{An object of class "dmm"}

\item{output_file}{(Optional) A file containing the vector of probabilities (e.g,"C:/.../PROB.txt")}

\item{plot}{`FALSE` (no figure plot of words probabilities); `TRUE` (figure plot)}
}
\value{
A dataframe of probability by position (and probability plots)
}
\description{
Probability of occurrence of the observed word of size n in a sequence at several positions
}
\examples{
\dontrun{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
n <- 2
length_probas(n, lambda, c(1,length(lambda)-n+1), dmm,plot=TRUE)
}
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{word_proba}
}
\author{
Victor Mataigne, Alexandre Seiller
}
