% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Loglikelihood}
\usage{
loglik(x, sequences)
}
\arguments{
\item{x}{An object of class "dmm" for which the logikelihood can be computed.}

\item{sequences}{A vector of character or list of vectors representing the sequences for which the
log-likelihood of the model must be computed.}

\item{states}{Vector of state space (of length s).}
}
\value{
A numeric, the log-likelihood
}
\description{
Generic function computing the loglikelihood of the model `x`,
  with the list of sequences `sequences`.
}
