% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{maintainability}
\alias{maintainability}
\title{Maintainability function}
\usage{
maintainability(x, k1, k2, s1, output_file = NULL, plot = FALSE)
}
\arguments{
\item{x}{An object of class `dmm`}

\item{k1}{Start position :  a positive integer giving the start position along the sequence from which the maintainabilities of the DMM should be computed}

\item{k2}{End position :  a positive integer giving the end position along the sequence until which the maintainabilities of the DMM should be computed}

\item{s1}{Character vector of the subspace working states among the state space vector such that s1 < s}

\item{output_file}{(Optional) A file containing matrix of maintainability probabilities (e.g, "C:/.../MAIN.txt")}

\item{plot}{`FALSE` (no figure plot of maintainability by position); `TRUE` (figure plot)}
}
\value{
A matrix with positions and maintainability probabilities of states (and figure plot)
}
\description{
Maintainability of a system at time \eqn{k \in N}.
}
\details{
The maintainability at time \eqn{k \in N} of a system is the probability that the system is repaired up to time \eqn{l},
given that is has failed at time \eqn{l=0}.
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
k1 <- 1
k2 <- 200
s1 <- c("c","t")  # vector of working states
maintainability(dmm,k1,k2,s1,plot=TRUE)
}
\references{
\insertRef{BaVe2018}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}
}
\author{
Alexandre Seiller
}
