% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{maintainability}
\alias{maintainability}
\title{Maintainability function}
\usage{
maintainability(x, k1, k2, upstates, output_file = NULL)
}
\arguments{
\item{x}{An object of class \code{dmm}}

\item{k1}{Start position :  a positive integer giving the start position along the sequence from which the maintainabilities of the DMM should be computed}

\item{k2}{End position :  a positive integer giving the end position along the sequence until which the maintainabilities of the DMM should be computed}

\item{upstates}{Character vector of the subspace working states among the state space vector such that upstates < s}

\item{output_file}{(Optional) A file containing matrix of maintainability probabilities (e.g, "C:/.../MAIN.txt")}
}
\value{
A vector of length k + 1 giving the values of the maintainability for the period \eqn{[0 \ldots k]}
}
\description{
Maintainability of a system at time \eqn{k \in N} is the probability that the system is repaired up to time \eqn{k},
given that is has failed at time \eqn{k=0}.
}
\details{
Consider a system (or a component) System whose possible states during its evolution in time are
\eqn{E = \{1 \ldots s \}}. Denote by \eqn{U = \{1 \ldots s_1 \}} the subset of operational states of the system (the upstates) and by \eqn{D =\{s_{1}+1 \ldots s \}} the subset of failure states (the down states), with 0 < s1 < s(obviously, \eqn{E = U \cup D and U \cap D = \emptyset, U \neq \emptyset, D \neq \emptyset}). One can think of the states of U as
different operating modes or performance levels of the system, whereas the states of D can be seen as failures of the systems with different modes.
}
\examples{

data(lambda, package = "drimmR")
dmm <- fitdmm(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq", fit.method="sum")
k1 <- 1
k2 <- 200
upstates <- c("c","t")  # vector of working states
getM <- maintainability(dmm,k1,k2,upstates)


# Figure plot example :

ggplot2::ggplot(data.frame(getM), aes(positions,maintainability)) + geom_path() +
theme_bw() + geom_point() +
 scale_y_continuous(name= "Maintainability") +
 scale_x_continuous(name= "Position",breaks = if(k2<=20){seq(from=0,to=k2, by=1)}
                    else if(k2>20 & k2<=100){seq(from=0,to=k2, by=10)}
                    else if(k2>100 & k2<=1000){seq(from=0,to=k2, by=100)}
                    else if(k2>1000 & k2<=10000){seq(from=0,to=k2, by=1000)}
                    else if(k2>10000 & k2<=100000){seq(from=0,to=k2, by=10000)}
                    else{seq(from=0,to=k2, by=100000)})


}
\references{
\insertRef{BaVe2018}{drimmR}
}
\seealso{
\link[drimmR]{fitdmm}, \link[drimmR]{getTransitionMatrix}
}
\author{
Alexandre Seiller
}
