% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{simulate.dmmsum}
\alias{simulate.dmmsum}
\title{Simulate a sequence with the Drifting Markov Model}
\usage{
\method{simulate}{dmmsum}(x, output_file, model_size = NULL)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{output_file}{File containing the simulated sequence}

\item{model_size}{Size of the model}
}
\value{
the vector of simulated sequence
}
\description{
Simulate a sequence with the Drifting Markov Model
}
\examples{
data(lambda, package = "drimmR")
SIM.out <- "C:\\\\...\\\\file.txt"
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
simulate(dmm,SIM.out,20000)

}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{getStationaryLaw}
}
\author{
Annthomy Gilles, Alexandre Seiller
}
