% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmsum.R
\name{simulate.dmmsum}
\alias{simulate.dmmsum}
\title{Simulate a sequence under a drifting Markov model}
\usage{
\method{simulate}{dmmsum}(x, output_file = NULL, model_size = NULL)
}
\arguments{
\item{x}{An object of class `dmm`, `dmmsum`}

\item{output_file}{(Optional) File containing the simulated sequence (e.g, "C:/.../SIM.txt")}

\item{model_size}{Size of the model}
}
\value{
the vector of simulated sequence
}
\description{
Simulate a sequence under a k-th order DMM.
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
simulate(dmm, model_size=100)
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{getStationaryLaw}
}
\author{
Annthomy Gilles, Alexandre Seiller
}
