% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{simulate}
\alias{simulate}
\title{Simulate a sequence under a drifting Markov model}
\usage{
simulate(x, output_file, model_size = 1e+05)
}
\arguments{
\item{x}{An object of class `dmm`}

\item{output_file}{(Optional) File containing the simulated sequence (e.g, "C:/.../SIM.txt")}

\item{model_size}{Size of the model}
}
\description{
Generic function simulating a sequence under a k-th order DMM.
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\author{
Annthomy Gilles, Alexandre Seiller
}
