% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{simulate}
\alias{simulate}
\title{Simulate S3 generic function}
\usage{
simulate(x, output_file, model_size = 1e+05)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{output_file}{A file containing matrix of probabilities}

\item{model_size}{Size of the model}
}
\description{
Simulate S3 generic function
}
\examples{
data(lambda, package = "drimmR")
SIM.out <- "C:\\\\...\\\\file.txt"
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
simulate(dmm,SIM.out,20000)

}
\author{
Annthomy Gilles, Alexandre Seiller
}
