% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{simulate}
\alias{simulate}
\title{Simulate S3 generic function}
\usage{
simulate(x, output_file, model_size = 1e+05)
}
\arguments{
\item{x}{An object of class "dmm", \link[drimmR]{dmmsum}}

\item{output_file}{(Optional) File containing the simulated sequence (e.g, "C:/.../SIM.txt")}

\item{model_size}{Size of the model}
}
\description{
Simulate S3 generic function
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\author{
Annthomy Gilles, Alexandre Seiller
}
