% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{stationaryLaw_evol}
\alias{stationaryLaw_evol}
\title{Plot stationary laws for a range of positions between <start> and <end>}
\usage{
stationaryLaw_evol(
  x,
  start = 1,
  end = NULL,
  step = NULL,
  output_file = NULL,
  plot = FALSE
)
}
\arguments{
\item{x}{An object of class "dmm"}

\item{start}{Start position (numeric)}

\item{end}{End position (numeric)}

\item{step}{A step (numeric)}

\item{output_file}{(Optional) A file containing matrix of stationary laws (e.g, "C:/.../SL.txt")}

\item{plot}{FALSE (no figure plot of SL evolution); TRUE (figure plot)}
}
\value{
A matrix of probabilities with position and probability of states (and figure plot)
}
\description{
Plot stationary laws for a range of positions between <start> and <end>
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
stationaryLaw_evol(dmm,start=1,end=1000,step=100, plot=TRUE)
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getStationaryLaw}
}
\author{
Alexandre Seiller
}
