% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word_expect}
\alias{word_expect}
\title{Expectation of a word in a DMM}
\usage{
word_expect(word, x, output_file)
}
\arguments{
\item{word}{A subsequence}

\item{DMM}{An object of class "dmm"}
}
\value{
a float
}
\description{
Expectation of a word in a DMM
}
\examples{
#' data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
PROB.out <- "C:\\\\...\\\\file.txt"
word_expect("atcggatc", dmm, output_file=PROB.out)
}
\author{
Victor Mataigne, Alexandre Seiller
}
