% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word_proba}
\alias{word_proba}
\title{Probability of a word at a position t of a DMM}
\usage{
word_proba(word, pos, x, output_file = NULL, internal = FALSE)
}
\arguments{
\item{word}{A subsequence (string of characters)}

\item{pos}{A position (numeric)}

\item{x}{An object of class "dmm"}

\item{output_file}{(Optional) A file containing the probability (e.g,"C:/.../PROB.txt")}

\item{internal}{FALSE (default) ; TRUE (for internal use of word applications)}
}
\value{
A numeric, probability of \code{word}
}
\description{
Probability of a word at a position t of a DMM
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
word_proba("aggctga",4,dmm)
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{word_probas}, \link[drimmR]{words_probas}
}
\author{
Victor Mataigne, Alexandre Seiller
}
