% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word_probas}
\alias{word_probas}
\title{Probabilities of a word at several positions of a DMM}
\usage{
word_probas(word, t, x, output_file = NULL)
}
\arguments{
\item{word}{A subsequence}

\item{t}{A vector of positions}

\item{x}{An object of class "dmm"}
}
\value{
A numeric vector, probabilities of \code{word}
}
\description{
Probabilities of a word at several positions of a DMM
}
\examples{
word_probas("aggctga",c(100,300),x)
}
\author{
Victor Mataigne
}
