% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word_probas}
\alias{word_probas}
\title{Probabilities of a word at several positions of a DMM}
\usage{
word_probas(word, pos, x, output_file = NULL, plot = FALSE)
}
\arguments{
\item{word}{A subsequence (string of characters)}

\item{pos}{A vector of integer positions}

\item{x}{An object of class "dmm"}

\item{output_file}{(Optional) A file containing the vector of probabilities (e.g,"C:/.../PROB.txt")}

\item{plot}{`FALSE` (no figure plot of word probabilities); `TRUE` (figure plot)}
}
\value{
A numeric vector, probabilities of \code{word}
}
\description{
Probabilities of a word at several positions of a DMM
}
\examples{
\dontrun{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
word_probas("aggctga",c(100,300),dmm, plot=TRUE)
}
}
\references{
\insertRef{BaVe2018}{drimmR}
\insertRef{Ver08}{drimmR}
}
\seealso{
\link[drimmR]{dmmsum}, \link[drimmR]{getTransitionMatrix}, \link[drimmR]{word_proba}, \link[drimmR]{words_probas}
}
\author{
Victor Mataigne, Alexandre Seiller
}
