% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{words_probas}
\alias{words_probas}
\title{Probability of several words at several positions of a DMM}
\usage{
words_probas(words, pos, x, output_file = NULL, plot = FALSE)
}
\arguments{
\item{words}{A vector of characters containing words}

\item{pos}{A vector of integer positions}

\item{x}{An object of class "dmm"}

\item{output_file}{A file containing the matrix of probabilities}

\item{plot}{FALSE (no figure plot of words probabilities); TRUE (figure plot)}
}
\value{
a dataframe
}
\description{
Probability of several words at several positions of a DMM
}
\examples{
data(lambda, package = "drimmR")
dmm <- dmmsum(lambda, 1, 1, c('a','c','g','t'), init.estim = "freq")
PROB.out <- "C:\\\\...\\\\file.txt"
words_probas(c("atcgattc", "taggct", "ggatcgg"),c(100,300),dmm, output_file=PROB.out, plot=FALSE)
}
\author{
Victor Mataigne, Alexandre Seiller
}
