% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{words_probas}
\alias{words_probas}
\title{Probability of several words at several positions of a DMM}
\usage{
words_probas(words, pos, x, output_file = NULL)
}
\arguments{
\item{words}{A vector, containing words}

\item{pos}{A vector of positions}

\item{x}{An object of class "dmm"}

\item{plot}{FALSE (no figure plot of words probabilities); TRUE (figure plot)}
}
\value{
a dataframe
}
\description{
Probability of several words at several positions of a DMM
}
\examples{
words_probas(c("atcgattc", "taggct", "ggatcgg"), c(100,200), x)
}
\author{
Victor Mataigne, Alexandre Seiller
}
